Compress added scheduling structure into existing owners
Fold the internal hook dispatch loop back into internal-hooks.ts
(bootstrap keeps opting into yield-between-handlers), retarget
bootstrap-hooks to the canonical entry point, drop the unused
handler-timing callback, deduplicate the bootstrap resolver parameter
types, and use the promisified close for the guarded workspace read.
Net -25 production lines with no behavior change; the event-loop
proof harness import is retargeted to the re-homed exports.
