+++ src/agents/bootstrap-hooks.ts
@@ -2,9 +2,11 @@
  * Applies internal agent bootstrap hooks before workspace context is injected.
  */
 import type { OpenClawConfig } from "../config/types.openclaw.js";
-import { triggerInternalHookWithScheduling } from "../hooks/internal-hook-dispatch.js";
 import type { AgentBootstrapHookContext } from "../hooks/internal-hooks.js";
-import { createInternalHookEvent } from "../hooks/internal-hooks.js";
+import {
+  createInternalHookEvent,
+  triggerInternalHookWithScheduling,
+} from "../hooks/internal-hooks.js";
 import { resolveAgentIdFromSessionKey } from "../routing/session-key.js";
 import type { WorkspaceBootstrapFile } from "./workspace.js";
